Attempt to Migrate `.bashrc` to Nix
Attempted to migrate stuff in `.bashrc` to Nix and failed miserably lol.
I DID migrate some aliases and shortened most if not all the comments in
`.bashrc` to make it cleaner. I may try this again soon but I dunno
because I have basically no experience in fiddling with `.bashrc` for
*reasons*.