Cleaning Flake
The flake was actually already pretty clean it looks great. I only just
added comments to describe each input of the flake like what it's for
and all that and made sure to use `inputs.` for referring to every
single flake input/dependency in the `outputs` block. And of course I
changed the flake's description just for funzies.